Split UI & panels. Renamed network to Entry point. Added additional engine scripts. UIMananger now working on device with Unity native UI.

+++ Assets/Engine/UI/UIRoot.cs
@@ -20,7 +20,7 @@ namespace FFEngine
 			
 			if(adaptToScreenRatio)
 			{
-				float ratio = Screen.width / Screen.height;
+				float ratio = (float)Screen.width / (float)Screen.height;
 				float multiplier = ratio * 3f / 4f;
 				SetWidthMultiplier(multiplier);
 			}
